requirements.txt
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
